Add CW MARS Customization
This commit adds code from KCLS to add the --for-text option to
create-notice-file.pl.

The KCLS --for-text code is modified to limit output of holds to only
those that have the sms_notify field set.

The default list of events and granularities to run are replaced with
those from CW MARS.

A helper script, run_hourly, is added to be run by cron on an hourly
basis to simplify processing the events.

Signed-off-by: Jason Stephenson <jstephenson@cwmars.org>